Player Movement: Collision Detection & Snapping
Player should collide and snap to surfaces without needing the Unity Character Controller or RigidBody (in this case, for 2D). It will detect up to two terrain collisions (using a mask) at the same time by entering the data into an array and using a For loop to process the normals of each collision.
WARNING: Not designed for use with slopes, so unintended interactions may occur with angled surfaces.