Enemy Spawner Script + Portals
There is now a fully dedicated script and object for managing enemy spawns and deleting enemies. It currently can spawn and modify values of portals to spawn any existing enemy type (feature work in progress, as there is only one enemy type right now). There is also a basic bat AI that chases the player in a straight line, but damage is not implemented yet.